puedo agregar un admin

+++ secretPal-lib/src/main/java/com/tenPines/application/service/UserService.java
@@ -15,9 +15,12 @@ public class UserService {
     private final UserRepository userRepository;
     private final JWTGenerator jwtGenerator;
 
-    public UserService(UserRepository userRepository, JWTGenerator jwtGenerator) {
+    private AdminService adminService;
+
+    public UserService(UserRepository userRepository, JWTGenerator jwtGenerator, AdminService adminService) {
         this.userRepository = userRepository;
         this.jwtGenerator = jwtGenerator;
+        this.adminService = adminService;
     }
 
     public User save(User aUser) {
@@ -25,11 +28,22 @@ public class UserService {
     }
 
     public void deleteByWorker(Worker worker) {
+        assertIsNotAdmin(worker);
         userRepository.deleteByWorker(worker);
+    }
+
+    private void assertIsNotAdmin(Worker worker) {
+        if(adminService.isAdmin(findByWorker(worker))){
+            throw new RuntimeException("No se puede borrar a un administrador");
+        }
     }
 
     public Optional<User> findByBackofficeId(Long backofficeId) {
         return userRepository.findByBackofficeId(backofficeId);
+    }
+
+    public User findByWorker(Worker worker) {
+        return userRepository.findByWorker(worker);
     }
 
     public String getTokenForUser(User user) {
